requirements.txt
sanic
aioredis
asynctest
pytest-sanic